services/crm/pyproject.toml
[project]
name = "crm"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "structlog>=24.4",
    "pydantic-settings>=2.0",
    "bss-models",
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "transitions>=0.9",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv.sources]
bss-models = { workspace = true }
